# Remove secret-dependant non-constant time operation in ecmult_const.