@sipa 2020-01-08
Obvious Concept ACK

Is there any measurable impact on performance?