@gmaxwell 2020-01-08
The project doesn't have any good way to benchmark really small changes, you just end up measuring noise from cache alignment of code and differences in branch predictor aliasing.

There has been some academic work on randomizing binary layout and running tests a bunch of times and averaging over them, but the implementation I found were unusably bitrotted. :(

But-- best effort,

Before:
ecdh: min 89.7us / avg 89.7us / max 90.0us
ecdsa_verify: min 77.7us / avg 77.7us / max 77.9us

After:
ecdh: min 88.0us / avg 88.0us / max 88.3us
ecdsa_verify: min 77.8us / avg 77.9us / max 77.9us

Wouldn't be shocking if there were a real improvement for those ECDHs, but I'd just assume that's noise.