@gmaxwell 2020-01-09
> this code gets faster on gcc and msvc if you make abs_n unsigned

I don't especially like peppering code with unsigned values where it's not needed (e.g. for overflow or range purposes), as it risks creating bugs due to  promotion and it prevents some optimizations that depend on static analysis being able to assume overflow won't happen.  ... in this case the only justification for the performance difference is / behaviour using >> 1 instead of /2 should result in the same behaviour as changing things to unsigned.

I also prefer the shift because on virtually every system an actual div instruction is extremely non-constant time. This means that the code is making a strong assumption that the compiler will perform strength reduction. I'm confident this isn't true and that sometimes they'll emit a div or a huge blob of division code on cpus without a divider arm (try -Os builds on older GCC).