@real-or-random 2020-01-10
ACK 9b3969c9d6076953faeaf9e96134ac2bab8ba7ee I read the diff carefully and tested the code with ECDH enabled and various settings, also on valgrind