@elichai 2020-01-15
ACK d567b779fe446fd18820a9d2968ecb703c8dea19 after merge. checked the asm of the abs/2 in godbolt on multiple different architectures and different compilers and non of them seem to use known variable time instructions.